requirements.txt
pygame==2.2.0
dynaconf==3.1.11
numpy==1.24.2
pandas==1.5.3
memory_profiler==0.61.0
scikit-learn==1.2.1
matplotlib==3.4.3
seaborn==0.11.2
umap-learn==0.5.1
dill==0.3.6
tqdm==4.62.3
pyfiglet==0.8.post1